backend/requirements.txt
pip>=23.0.1  # Para asegurar una versión actualizada de pip
setuptools>=69.0.0  # Para pkg_resources y otras utilidades
Django>=4.2.0,<4.3.0
djangorestframework>=3.14.0,<3.15.0
django-cors-headers>=4.1.0,<4.2.0
python-dotenv>=1.0.0
Pillow>=10.0.0  # Para manejo de imágenes
django-filter==23.5  # Para filtros avanzados en DRF
drf-spectacular>=0.26.2,<0.27.0  # Para documentación de API
channels>=4.0.0,<4.1.0  # Para mensajería en tiempo real
psycopg2-binary>=2.9.6,<2.10.0  # Para PostgreSQL en producción
pytest>=7.0.0  # Para testing
pytest-django>=4.5.0
python-jose>=3.3.0  # Para manejo de JWT
djangorestframework-simplejwt>=5.2.2,<5.3.0  # Para autenticación JWT
django-environ>=0.10.0,<0.11.0
gunicorn==21.2.0
cloudinary==1.36.0
django-cloudinary-storage==0.3.0
whitenoise==5.2.0
qrcode>=0.18.0

